#129 修正

+++ api/app/controllers/api/v1/password_reset_controller.rb
@@ -2,25 +2,25 @@ require "securerandom"
 class Api::V1::PasswordResetController < ApplicationController
   # パスワードリセット
   def create
-    # 指定したIDのパスワードリセットトークンをすべて削除
-    password_reset_tokens = PasswordResetToken.where(admin_id: params[:id])
-    password_reset_tokens.destroy_all
-
-    # トークンの生成
-    token = PasswordResetToken.new(token: generate_token, admin_id: params[:id])
-    token.save!
-
     # 管理者の指定
     admin = Admin.find_by(id: params[:id])
 
-    url = format(URL_ADMIN_PASSWORD_RESET_TOKEN, token: token.token)
-    mail = PasswordResetMailer.creation_email(admin, url).deliver_now
+    if admin.is_invalid == false
+      # 指定したIDのパスワードリセットトークンをすべて削除
+      password_reset_tokens = PasswordResetToken.where(admin_id: params[:id])
+      password_reset_tokens.destroy_all
 
-    render json: json_render_v1(true)
-  end
+      # トークンの生成
+      token = PasswordResetToken.new(token: generate_token, admin_id: params[:id])
+      token.save!
 
-  # トークンの生成
-  def generate_token
-    SecureRandom.urlsafe_base64
+      url = format(URL_ADMIN_PASSWORD_RESET_TOKEN, token: token.token)
+      mail = PasswordResetMailer.creation_email(admin, url).deliver_now
+
+      render json: json_render_v1(true)
+    else
+      render json: json_render_v1(false)
+    end
+    nil
   end
 end
